+++ src/test/java/org/folio/processing/mapping/MappingTest.java
@@ -40,9 +40,11 @@ public class MappingTest {
   private static final String PRECEDING_FILE_PATH = "src/test/resources/org/folio/processing/mapping/780_785_examples.mrc";
   private static final String BIBS_ERRORS_PATH = "src/test/resources/org/folio/processing/mapping/test1_err.mrc";
   private static final String BIB_WITH_REPEATED_SUBFIELDS_PATH = "src/test/resources/org/folio/processing/mapping/336_repeated_subfields.mrc";
-  private static final String BIB_WITH_880_with_111_subfield_value = "src/test/resources/org/folio/processing/mapping/880_111_to_711.mrc";
-  private static final String BIB_WITH_880_2_with_245_subfield_value = "src/test/resources/org/folio/processing/mapping/880_245_to_246.mrc";
-  private static final String BIB_WITH_880_3_with_830_subfield_value = "src/test/resources/org/folio/processing/mapping/880_to_830.mrc";
+  private static final String BIB_WITH_880_WITH_111_SUBFIELD_VALUE = "src/test/resources/org/folio/processing/mapping/880_111_to_711.mrc";
+  private static final String BIB_WITH_880_2_WITH_245_SUBFIELD_VALUE = "src/test/resources/org/folio/processing/mapping/880_245_to_246.mrc";
+  private static final String BIB_WITH_880_3_WITH_830_SUBFIELD_VALUE = "src/test/resources/org/folio/processing/mapping/880_to_830.mrc";
+  private static final String BIB_WITH_5xx_STAFF_ONLY_INDICATORS = "src/test/resources/org/folio/processing/mapping/5xx_staff_only_indicators.mrc";
+
   private static final String DEFAULT_MAPPING_RULES_PATH = "src/test/resources/org/folio/processing/mapping/rules.json";
   private static final String STUB_FIELD_TYPE_ID = "fe19bae4-da28-472b-be90-d442e2428ead";
   private static final String BIB_WITH_MISSING_URI = "src/test/resources/org/folio/processing/mapping/856_missing_uri.mrc";
@@ -121,7 +123,7 @@ public class MappingTest {
 
   @Test
   public void testMarcToInstance880FieldToContributorMeetingName() throws IOException {
-    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_880_with_111_subfield_value).getBytes(StandardCharsets.UTF_8)));
+    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_880_WITH_111_SUBFIELD_VALUE).getBytes(StandardCharsets.UTF_8)));
     JsonObject mappingRules = new JsonObject(TestUtil.readFileFromPath(DEFAULT_MAPPING_RULES_PATH));
 
     ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
@@ -148,7 +150,7 @@ public class MappingTest {
 
   @Test
   public void testMarcToInstance880FieldToAlternativeTitleName() throws IOException {
-    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_880_2_with_245_subfield_value).getBytes(StandardCharsets.UTF_8)));
+    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_880_2_WITH_245_SUBFIELD_VALUE).getBytes(StandardCharsets.UTF_8)));
     JsonObject mappingRules = new JsonObject(TestUtil.readFileFromPath(DEFAULT_MAPPING_RULES_PATH));
 
     ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
@@ -172,7 +174,7 @@ public class MappingTest {
 
   @Test
   public void testMarcToInstance880FieldToSeriesStatement() throws IOException {
-    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_880_3_with_830_subfield_value).getBytes(StandardCharsets.UTF_8)));
+    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_880_3_WITH_830_SUBFIELD_VALUE).getBytes(StandardCharsets.UTF_8)));
     JsonObject mappingRules = new JsonObject(TestUtil.readFileFromPath(DEFAULT_MAPPING_RULES_PATH));
 
     ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
@@ -189,6 +191,33 @@ public class MappingTest {
       Assert.assertNotNull(instance.getSeries());
       Assert.assertEquals(1, instance.getSeries().size());
       Assert.assertNotNull(instance.getSeries().stream().filter(e -> e.equals("testingSeries")).findAny().orElse(null));
+      Validator validator = factory.getValidator();
+      Set<ConstraintViolation<Instance>> violations = validator.validate(instance);
+      Assert.assertTrue(violations.isEmpty());
+    }
+  }
+
+  @Test
+  public void testMarcToInstanceNoteStaffOnlyViaIndicator() throws IOException {
+    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_5xx_STAFF_ONLY_INDICATORS).getBytes(StandardCharsets.UTF_8)));
+    JsonObject mappingRules = new JsonObject(TestUtil.readFileFromPath(DEFAULT_MAPPING_RULES_PATH));
+
+    ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+    while (reader.hasNext()) {
+      ByteArrayOutputStream os = new ByteArrayOutputStream();
+      MarcJsonWriter writer = new MarcJsonWriter(os);
+      Record record = reader.next();
+      writer.write(record);
+      JsonObject marc = new JsonObject(new String(os.toByteArray()));
+      Instance instance = mapper.mapRecord(marc, new MappingParameters(), mappingRules);
+      Assert.assertNotNull(instance.getTitle());
+      Assert.assertNotNull(instance.getSource());
+      Assert.assertNotNull(instance.getNotes());
+      Assert.assertEquals(3, instance.getNotes().size());
+      Assert.assertEquals("Rare copy: Gift of David Pescovitz and Timothy Daly", instance.getNotes().get(1).getNote());
+      Assert.assertTrue( instance.getNotes().get(1).getStaffOnly());
+      Assert.assertEquals("Correspondence relating to the collection may be found in Cornell University Libraries. John M. Echols Collection. Records, #13\\\\6\\\\1973", instance.getNotes().get(2).getNote());
+      Assert.assertFalse( instance.getNotes().get(2).getStaffOnly());
       Validator validator = factory.getValidator();
       Set<ConstraintViolation<Instance>> violations = validator.validate(instance);
       Assert.assertTrue(violations.isEmpty());